convert Double to Float

+++ app/src/main/java/com/example/helloworldkotlinandroid/MainActivity.kt
@@ -72,8 +72,8 @@ class MainActivity : AppCompatActivity(), LocationListener {
 
                 // Pass the resolved celestial path arrays straight to the matrix transformer
                 celestialCalibrator.performCelestialCalibration(
-                    moonTarget.azimuth,
-                    moonTarget.altitude,
+                    moonTarget.azimuth.toFloat(),
+                    moonTarget.altitude.toFloat(),
                 )
 
                 val telemetryReport =
